Update business entities to simple java objects
