+++ BusinessEntitiesAPI/src/main/java/businessentitiesapi/Customer.java
@@ -6,21 +6,29 @@ package businessentitiesapi;
 import java.time.LocalDate;
 
 /**
- * Customer class, representing the customer of a company.
+ *
  * @author Richard van den Ham {@code r.vandenham@fontys.nl}
  */
-public interface Customer {
+public class Customer{
     
-    /**
-     * Get name of customer.
-     * @return customer name
-     */
-    String getName();   
-    
-    /**
-     * Get date of birth of customer.
-     * @return date of birth
-     */
-    LocalDate getDateOfBirth();
-    
+    private final String name;
+    private final LocalDate dateOfBirth;
+
+    public Customer( String name, LocalDate dob ) {
+        this.name = name;
+        this.dateOfBirth = dob;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalDate getDateOfBirth() {
+        return dateOfBirth;
+    }
+
+    @Override
+    public String toString() {
+        return "Customer{" + "name=" + name + ", dateOfBirth=" + dateOfBirth + '}';
+    }
 }
