+++ Persistence/src/main/java/persistence/CustomerStorageServiceImpl.java
@@ -4,7 +4,6 @@
 package persistence;
 
 import businessentitiesapi.Customer;
-import businessentitiesapi.CustomerManager;
 import java.time.LocalDate;
 import java.time.Month;
 import java.util.ArrayList;
@@ -18,17 +17,7 @@ import java.util.List;
  * @author Richard van den Ham {@code r.vandenham@fontys.nl}
  */
 class CustomerStorageServiceImpl implements CustomerStorageService{
-
-    /**
-     * The customerManager can be used to create Customer objects, 
-     * which is delegated to the business logic layer.
-     */
-    private final CustomerManager customerManager;
-
-    CustomerStorageServiceImpl(CustomerManager customerManager) {
-        this.customerManager = customerManager;
-    }
-
+    
     @Override
     public void add( Customer c ) {
         System.out.println( "DO SMART THINGS TO STORE A CUSTOMER" );
@@ -41,8 +30,8 @@ class CustomerStorageServiceImpl implements CustomerStorageService{
         // for each entry in resultset, make customerobject using factory method in CustomerManager.
         // For now, simulate DB result:
         
-        Customer dummy1 = customerManager.createCustomer("Lisa", LocalDate.of(1980, 8, 1));
-        Customer dummy2 = customerManager.createCustomer("Jack", LocalDate.of(1960, 1, 8));
+        Customer dummy1 = new Customer("Lisa", LocalDate.of(1980, 8, 1));
+        Customer dummy2 = new Customer("Jack", LocalDate.of(1960, 1, 8));
         
         return new ArrayList<>( Arrays.asList( dummy1, dummy2 ));
